Refactor: Extract page components and add metadata
Extract page logic into reusable components and add metadata for SEO.
